+++ image-loader/app/model/ImageUpload.scala
@@ -96,6 +96,7 @@ case object ImageUpload {
       metadata,
       usageRights,
       usageRights,
+      List(),
       List()
     )
   }
